[front] 개인 경로 찾을 시 데이터 저장 + 페인 페이지 분리
